requirements.txt
Twisted==14.0.2
blist==1.3.6
cassandra-driver==2.1.2
cqlengine==0.20.0
